Backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
requests
beautifulsoup4
python-dotenv
pydantic
google-generativeai==0.8.5
